Github Issue #50: accounts::group class does not exist
When using puppetlabs/puppetlabs-accounts the example files provide "templates" for using the class and adding a user and groups.

If you follow the current examples they will fail for two reasons:

* accounts::group does not exist
* the examples include non-quoted variables which fail the validate_re checks

I have added manifests/group.pp to address the missing class and updated the example files to quote the gid so that validate_re treats it as a string.

I've tested locally and this fixes the issue for me and I have successfully added a user and a separate group with this module now.

I haven't updated the spec tests as I am not that familiar with that part of Puppet (yet). :D

I believe this PR will also fix https://tickets.puppetlabs.com/browse/MODULES-3114

Any questions let me know!

Cheers,

Amo
@sysadmiral